refactor: disposed the idea of having a bridge and removed the singleton, now the data is on a DataStore
